logic-for-btn-in-modal-window

+++ src/js/functions/local-storage.js
@@ -37,6 +37,37 @@ let id;
 
 export function modalBtnsHandler(externalId) {
   id = Number(externalId);
+
+  if (localStorage.getItem('userID') !== '' && localStorage.getItem('userID') !== null){
+      let filmInListWathed = 0;
+      listOfDataWathed.map(key => {
+      if(Number(key.id) === id){return filmInListWathed = id}});
+      console.log('id', filmInListWathed);
+      if(filmInListWathed === 0){
+        watchedBtnRef.textContent = 'ADD TO WATCHED';
+        watchedBtnRef.classList.remove('modal__button--active');
+      } else {
+        watchedBtnRef.textContent = 'REMOVE FROM WATCHED';
+        watchedBtnRef.classList.add('modal__button--active');
+      }
+
+      let filmInListQueue = 0;
+      listOfDataQueue.map(key => {
+      if(Number(key.id) === id){return filmInListQueue = id}});
+      console.log('id', filmInListQueue);
+      if(filmInListQueue === 0){
+        queueBtnRef.textContent = 'ADD TO QUEUE';
+        queueBtnRef.classList.remove('modal__button--active');
+      } else {
+        queueBtnRef.textContent = 'REMOVE FROM QUEUE';
+        queueBtnRef.classList.add('modal__button--active');
+      }
+  }
+  else 
+  { if(!localStorage.getItem('watched')) {
+    watchedBtnRef.textContent = 'ADD TO WATCHED';
+    watchedBtnRef.classList.remove('modal__button--active');
+
   const user = localStorage.getItem('statusUser');
 
   if (user === 'anonym') {
@@ -45,12 +76,21 @@ export function modalBtnsHandler(externalId) {
       watchedBtnRef.textContent = 'ADD TO WATCHED';
       watchedBtnRef.classList.remove('modal__button--active');
 
+
       watched = [];
       localStorage.setItem('watched', JSON.stringify(watched));
     }
     if (!localStorage.getItem('queue')) {
       queueBtnRef.textContent = 'ADD TO QUEUE';
       queueBtnRef.classList.remove('modal__button--active');
+
+
+    queue = [];
+    localStorage.setItem('queue', JSON.stringify(queue));
+  }
+  
+  watched = JSON.parse(localStorage.getItem('watched'));
+  isAddedToWatched = watched.findIndex(film => Number(film.id) === Number(id));
 
       queue = [];
       localStorage.setItem('queue', JSON.stringify(queue));
@@ -60,6 +100,7 @@ export function modalBtnsHandler(externalId) {
     isAddedToWatched = watched.findIndex(
       film => Number(film.id) === Number(id)
     );
+
 
     if (isAddedToWatched === -1) {
       watchedBtnRef.textContent = 'ADD TO WATCHED';
@@ -83,6 +124,24 @@ export function modalBtnsHandler(externalId) {
     watchedBtnRef.addEventListener('click', onWatchedBtnClick);
     queueBtnRef.addEventListener('click', onQueueBtnClick);
   } else {
+
+    queueBtnRef.textContent = 'REMOVE FROM QUEUE';
+    queueBtnRef.classList.add('modal__button--active');
+  }
+}
+  watchedBtnRef.addEventListener('click', onWatchedBtnClick);
+  queueBtnRef.addEventListener('click', onQueueBtnClick);
+}
+
+export function onWatchedBtnClick() {
+  let watched = [];
+  localStorage.getItem('userID') !== '' && localStorage.getItem('userID') !== null ? watched = listOfDataWathed : watched = JSON.parse(localStorage.getItem('watched'));
+  const isAddedToWatched = watched.findIndex(film => Number(film.id) === id);
+  let watchedFilms;
+
+  if (isAddedToWatched === -1) {
+    if (localStorage.getItem('userID') !== '' && localStorage.getItem('userID') !== null){
+
     watched = listOfDataWathed;
     console.log(watched);
     isAddedToWatched = watched.findIndex(
@@ -140,6 +199,7 @@ export function onWatchedBtnClick() {
     if (isAddedToWatched === -1) {
       watchedFilms = [...watched, filmData.data];
       console.log(filmData.data);
+
       writeUserDataWathed(
         localStorage.getItem('userID'),
         filmData.data.id,
@@ -156,6 +216,34 @@ export function onWatchedBtnClick() {
       watchedBtnRef.textContent = 'REMOVE FROM WATCHED';
       watchedBtnRef.classList.add('modal__button--active');
     } else {
+
+    watchedFilms = [...watched, filmData.data];
+    console.log(filmData.data);
+    localStorage.setItem('watched', JSON.stringify(watchedFilms));
+    watchedBtnRef.textContent = 'REMOVE FROM WATCHED';
+    watchedBtnRef.classList.add('modal__button--active');
+  }} else {
+    if (localStorage.getItem('userID') !== '' && localStorage.getItem('userID') !== null){
+      deleteUserDataWathed(localStorage.getItem('userID'), filmData.data.id);
+      watchedBtnRef.textContent = 'ADD TO WATCHED';
+      watchedBtnRef.classList.remove('modal__button--active');
+    } else {
+    watchedFilms = watched.filter(movie => movie.id !== id);
+    localStorage.setItem('watched', JSON.stringify(watchedFilms));
+    watchedBtnRef.textContent = 'ADD TO WATCHED';
+    watchedBtnRef.classList.remove('modal__button--active');
+  }}
+}
+
+export function onQueueBtnClick() {
+  let queue = [];
+  localStorage.getItem('userID') !== '' && localStorage.getItem('userID') !== null ? queue = listOfDataQueue: queue = JSON.parse(localStorage.getItem('queue'));
+  const isAddedToQueue = queue.findIndex(film => Number(film.id) === id);
+  let queueFilms;
+
+  if (isAddedToQueue === -1) {
+    if (localStorage.getItem('userID') !== '' && localStorage.getItem('userID') !== null){
+
       watchedFilms = watched.filter(movie => movie.id !== id);
       deleteUserDataWathed(localStorage.getItem('userID'), filmData.data.id);
       watchedBtnRef.textContent = 'ADD TO WATCHED';
@@ -189,6 +277,7 @@ export function onQueueBtnClick() {
     let queueFilms;
     if (isAddedToQueue === -1) {
       queueFilms = [...queue, filmData.data];
+
       writeUserDataQueue(
         localStorage.getItem('userID'),
         filmData.data.id,
@@ -205,12 +294,31 @@ export function onQueueBtnClick() {
       queueBtnRef.textContent = 'REMOVE FROM QUEUE';
       queueBtnRef.classList.add('modal__button--active');
     } else {
+
+    queueFilms = [...queue, filmData.data];
+    localStorage.setItem('queue', JSON.stringify(queueFilms));
+    queueBtnRef.textContent = 'REMOVE FROM QUEUE';
+    queueBtnRef.classList.add('modal__button--active');
+  } }
+  else {
+    if (localStorage.getItem('userID') !== '' && localStorage.getItem('userID') !== null){
+      deleteUserDataQueue(localStorage.getItem('userID'), filmData.data.id);
+      queueBtnRef.textContent = 'ADD TO QUEUE';
+      queueBtnRef.classList.remove('modal__button--active');
+    } else {
+    queueFilms = queue.filter(movie => movie.id !== id);
+    localStorage.setItem('queue', JSON.stringify(queueFilms));
+    queueBtnRef.textContent = 'ADD TO QUEUE';
+    queueBtnRef.classList.remove('modal__button--active');
+  }}
+
       queueFilms = queue.filter(movie => movie.id !== id);
       deleteUserDataQueue(localStorage.getItem('userID'), filmData.data.id);
       queueBtnRef.textContent = 'ADD TO QUEUE';
       queueBtnRef.classList.remove('modal__button--active');
     }
   }
+
 }
 
 export function getWatchedItems() {
